Build my own TOC sidebar
Hook in to the sphinx event to add data to the template context,
and add some table-of-contents details for the current page.
This relies on the env giving us a table of contents that only
includes top-level headings for the current page (or at least
that all of my pages only include one level of headings).

Change-Id: Ic7d2b8433e8f63e16ef235ee5acc0c32c793fcfb