readline: gnureadline; log format; manual verification
Add gnureadline to all of the examples, since the standard downloadable
installer on OS X does not include readline support.

Change the log file format to only include the interesting messages and
not the level or logger name.

Manually re-run all examples and verify the output, with the new log
format.