set up separate book build
Several sections need to be treated as appendices for the book, so they
won't show up in the toctree. Since the 'only' and 'ifconfig' directives
don't work for the LaTeX builder, we need a completely separate root
document. Switching documents requires managing the exclusion list to
avoid warnings when we ignore that file while building the HTML.