pearson: fix xref handling
Because the xref handling assumes that all references will be between
files that are part of the doctree, and we're building separate doctrees
for each chapter, we need to first process *all* of the documents and
save the domain data that results so we can stick it back into place for
the xref machinery to use.

This is only done for the 'std' domain, but may also need to be done for
the other domains.